[refact] keyboard automata refactoring #36
기능 분리를 통해 모듈화
