+++ CustomKeyboard/Hangul/KeyboardIOManager.swift
@@ -1,0 +1,12 @@
+//
+//  KeyboardIOManager.swift
+//  CustomKeyboard
+//
+//  Created by 신의연 on 2022/07/19.
+//
+
+import Foundation
+
+class KeyboardIOManager {
+    
+}
